move program to back to main window prescription
